Use font-size for cn-magnification, fixes size of grace note heads
For cn-magnification, use the magstep of font-size instead of the
ratio of actual staff-space to cn-base-staff-space. This works
because we are now using grob callbacks rather than engravers.
The previous approach was for working at the engraver / context
stage, with context properties (cnBaseStaffSpace) but at the grob
callback stage we can just use the font-size of grobs.

This also fixes the size of grace note heads (acciaccatura,
appoggiatura, etc.), which LilyPond sets via the font-size of
NoteHead grobs. Previously grace note heads were the same size
as regular note heads.