Use a custom accidental style rather than an accidental engraver
With Clairnote's chromatic staff, accidentals are printed for
different notes than on a traditional staff.  Usually fewer
accidentals are needed, but in rare cases Clairnote prints an
accidental that wouldn't be there in traditional notation
(e.g. E-sharp followed by F-natural).

The previous solution was to use the dodecaphonic-no-repeat
accidental style to generate accidental grobs for almost all
notes, and then delete those that weren't needed (most of them)
with a custom accidental engraver.

Instead, we now use a custom accidental style so that only the
accidental grobs that are needed are created.  Then an engraver
is not needed.