Update accidental style code for LilyPond 2.24
Now measure-position is no longer passed to the accidental style
function, and alterations have end-moment instead of measure-position.

One change this entails is that in the custom cnAlterations context
property we now just store the alter value instead of the full
alteration definition. The only thing we ever usedkfrom cnAlterations
was the alter value.